Add a settings option to allow arbitrary HTML to be used in info pages
We've had the complaint from contributors to the Pombola site in South
Africa (People's Assembly) that the current admin interface for entering
blog posts doesn't allow them to include the kind of visualizations
they'd like to include on the site - in particular, they wanted to be
able to include either <script> or <iframe> elements. Currently the
django-markitup widget that we use will strip these out.

I'm obviously reluctant to allow these elements because of security
concerns: if someone malicious has access to the admin interface (or we
later allow blog or info pages to be created by other means) then the
malicious person could trivially create XSS (cross-site scripting)
attacks.

However, so long as access to the admin interface is restricted to
trusted administrators allowing arbitrary HTML should be OK.

This commit introduces a new option in the Django settings
(INFO_PAGES_ALLOW_RAW_HTML) which is set to False by default, but set to
True for South Africa.  When this option is enabled, then the editing
interface for info pages / blog posts has a "Use raw HTML?" checkbox,
which when selected, switches the Markdown entry box for a plain
textarea for HTML.

<script> tags are filtered out on every page except those that display a
single info page / blog post, so that they won't conflict with other
such Javascript, or run into problems operating on a truncated version
of the content.

Fixes #1553 and #1501.