Don't convert empty tags to self-closing tags in raw HTML blog posts
There was a problem that content after embedded videos (included via
YouTube's suggested <iframe ...></iframe> HTML) wasn't being displayed.
This turned out to be because we pass the raw HTML through lxml to
ensure that the output is well-formed, but this was converting empty
tags to self-closing tags and <iframe .../> and that eats the following
content in the page.

Adding method="html" to lxml.etree's tostring function preserves the
separate close tag.

Fixes #1608