Remove chevron-{down,up}-alt.png and associated css, img, and js
chevron-down-alt.png was used in two places: first on
`h2 a.appearances-trigger` (which doesn’t seem to be used
in any templates any more) and also in mobile-functions.js
which adds a button if a `.page-title` element has a
`data-sub-menu-id` attribute. The attribute is only
present in the default infopage_detail.html template,
but that isn’t used by any views, and is only extended by
south_africa/info_newsletter.html which overrides the `content`
block anyway to remove the `data-sub-menu-id`.

Seems like the whole lot is legacy cruft that can just be
removed. So that’s what I’m doing.